WIP - split out http uploader
