+++ src/datasink/kafka.rs
@@ -100,7 +100,7 @@ impl KafkaSink {
                     "curl".to_string()
                 };
                 thread::spawn(move || {
-                    http_producer(
+                    kafka_http_producer(
                         cutoff,
                         &curl_cmd,
                         &rest_endpoint,
@@ -384,9 +384,9 @@ fn make_sender_adapter(
 // environment.  The main thread does not wait for it to finish but spins up threads to handle its
 // stdin/stdout/stderr and the final wait.
 
-fn http_producer(
+fn kafka_http_producer(
     cutoff: Option<usize>,
-    cmd: &str,
+    curl_cmd: &str,
     api_endpoint: &str,
     http_proxy: &str,
     client_id: &str,
@@ -396,40 +396,68 @@ fn http_producer(
     incoming_message_queue: channel::Receiver<Message>,
     control_and_errors: channel::Sender<Operation>,
 ) {
-    let op = HttpBackgroundProducer {
-        cmd,
+    let uploader = HttpUploader::new(
+        curl_cmd,
         api_endpoint,
         http_proxy,
+        timeout,
+    );
+    let op = KafkaHttpBackgroundProducer {
+        uploader,
         client_id,
         sasl_identity: &sasl_identity,
-        timeout,
         control_and_errors,
     };
     background_producer(cutoff, sending_window, incoming_message_queue, &op);
 }
 
-struct HttpBackgroundProducer<'a> {
-    cmd: &'a str,
-    api_endpoint: &'a str,
-    http_proxy: &'a str,
+struct KafkaHttpBackgroundProducer<'a> {
+    uploader: HttpUploader,
     client_id: &'a str,
     sasl_identity: &'a Option<(String, String)>,
-    timeout: u64,
     control_and_errors: channel::Sender<Operation>,
 }
 
-impl<'a> BackgroundSender for HttpBackgroundProducer<'a> {
+impl<'a> BackgroundSender for KafkaHttpBackgroundProducer<'a> {
     fn send_all(&self, _id: usize, backlog: Vec<Msg>) {
-        http_send_messages(
-            &self.cmd,
-            &self.api_endpoint,
-            &self.http_proxy,
-            &self.client_id,
-            &self.sasl_identity,
-            self.timeout,
-            &self.control_and_errors,
-            backlog,
-        );
+        match self.uploader.start() {
+            Ok(stream) => {
+                let cred = if let Some((user, pass)) = &self.sasl_identity {
+                    format!("\"sasl-user\":\"{user}\", \"sasl-password\":\"{pass}\",")
+                } else {
+                    "".to_string()
+                };
+                let client = self.client_id.to_string();
+                for Msg {
+                    topic,
+                    key,
+                    value,
+                    ..
+                } in backlog
+                {
+                    let data_size = value.len();
+                    let ctrl = format!(
+                        "\n{{\"topic\":\"{topic}\",\"key\":\"{key}\",\"client\":\"{client}\",{cred}\"data-size\":{data_size}}}\n"
+                    );
+                    stream.put(ctrl.as_bytes());
+                    stream.put(value.as_bytes());
+                }
+                // This catches synchronous errors.  For async errors we're going to need a callback.
+                // Possibly the callback is a parameter to end().
+                match stream.end() {
+                    Ok(()) => {}
+                    Err(e) => {
+                        let _ = self.control_and_errors
+                            .send(Operation::Error(format!("Failed to send some data: {:?}", e)));
+                    }
+                }
+            }
+            Err(e) => {
+                // Not found, permission denied, etc.
+                let _ = self.control_and_errors
+                    .send(Operation::Fatal(format!("Failed to start uploader: {:?}", e)));
+            }
+        }
     }
 
     fn shutdown_delay_ms(&self) -> usize {
@@ -438,160 +466,6 @@ impl<'a> BackgroundSender for HttpBackgroundProducer<'a> {
 
     fn metadata_size(&self) -> (usize, usize) {
         // Conservative overhead for punctuation, field names, etc of the control object
-        (10 /* per batch */, 100 /* per message */)
-    }
-}
-
-// The abstraction here will be a "http poster" that encapsulates the sending method and a bunch of
-// the other stuff, and into which we can pump a bunch of data.  Then it does not matter if that
-// uses curl or some other method.  The http poster can be shared between the kafka code and the
-// pure http POST code, and has the same general shape as the kafka poster code.
-
-fn http_send_messages(
-    cmd_name: &str,
-    api_endpoint: &str,
-    http_proxy: &str,
-    client_id: &str,
-    sasl_identity: &Option<(String, String)>,
-    mut timeout: u64,
-    control_and_errors: &channel::Sender<Operation>,
-    msgs: Vec<Msg>,
-) {
-    // Curl will retry for 1s, 2s, 4s, ..., 10m and then stick to 10m
-    let mut retry_count = 0;
-    let mut next = 1;
-    while timeout > 0 {
-        timeout -= min(next, timeout);
-        next = min(600, next * 2);
-        retry_count += 1;
-    }
-    let mut args = vec![
-        "--silent".to_string(),
-        "--show-error".to_string(),
-        "--data-binary".to_string(),
-        "@-".to_string(),
-        "-H".to_string(),
-        "Content-Type: application/octet-stream".to_string(),
-    ];
-    if retry_count > 0 {
-        args.push("--retry".to_string());
-        args.push(format!("{}", retry_count));
-        args.push("--retry-connrefused".to_string());
-    }
-    args.push(api_endpoint.to_string());
-
-    // Really want to merge stdout and stderr
-    let mut cmd = std::process::Command::new(cmd_name);
-    log::debug!("Curl: {cmd_name} {:?}", args);
-    cmd.args(args)
-        .stdin(std::process::Stdio::piped())
-        .stdout(std::process::Stdio::piped())
-        .stderr(std::process::Stdio::piped());
-    if http_proxy != "" {
-        cmd.env("http_proxy", http_proxy)
-            .env("https_proxy", http_proxy);
-    }
-    match cmd.spawn() {
-        Ok(mut child) => {
-            if let (Some(mut stdin), Some(mut stdout), Some(mut stderr)) =
-                (child.stdin.take(), child.stdout.take(), child.stderr.take())
-            {
-                let cred = if let Some((user, pass)) = sasl_identity {
-                    format!("\"sasl-user\":\"{user}\", \"sasl-password\":\"{pass}\",")
-                } else {
-                    "".to_string()
-                };
-                let client = client_id.to_string();
-                drop(std::thread::spawn(move || {
-                    for Msg {
-                        timestamp,
-                        topic,
-                        key,
-                        value,
-                    } in msgs
-                    {
-                        let _ = timestamp;
-                        let data_size = value.len();
-                        let ctrl = format!(
-                            "\n{{\"topic\":\"{topic}\",\"key\":\"{key}\",\"client\":\"{client}\",{cred}\"data-size\":{data_size}}}\n"
-                        );
-                        if let Err(err) = stdin.write_all(ctrl.as_bytes()) {
-                            log::debug!("Failed to write control object: {:?}", err);
-                        }
-                        if let Err(err) = stdin.write_all(value.as_bytes()) {
-                            log::debug!("Failed to write data blob size {data_size}: {:?}", err);
-                        }
-                    }
-                    // Does this happen too soon?  As in, if there's enough data, will the consumer
-                    // not have finished consuming?  Or will it just flush things and all will be fine?
-                    drop(stdin);
-                }));
-                // Separate threads do the output consuming and waiting for curl to finish, in order to
-                // guarantee several things:
-                //
-                //  - if the curl does not terminate immediately (because it is retrying, not
-                //    reaching the host, bandwidth-limited, ...) then Sonar does not hang waiting
-                //    for it, but can get on with its work.
-                //  - curl will not block writing its output on a full pipe
-                //  - sonar will not block on writing to curl because curl is blocked
-                //  - the child does not linger once it's ready to exit
-                //
-                // We can't use wait_with_output() because that will close stdin, and we can't wait
-                // to fork off these thread until writing has completed.  We could maybe combine the
-                // two reader threads into one using some kind of nonblocking I/O.  Maybe there are
-                // other tricks.
-                //
-                // 1K buffer is enough to see interesting errors.
-                drop(std::thread::spawn(move || {
-                    let mut buf = [0; 1024];
-                    loop {
-                        match stdout.read(&mut buf[..]) {
-                            Err(_) | Ok(0) => {
-                                break;
-                            }
-                            Ok(n) => {
-                                // This is the common case even when curl fails to deliver when
-                                // the server rejects the message.
-                                log::debug!(
-                                    "Curl succeeded with output: {}",
-                                    String::from_utf8_lossy(&buf[..n])
-                                );
-                            }
-                        }
-                    }
-                    drop(stdout);
-                }));
-                drop(std::thread::spawn(move || {
-                    let mut buf = [0; 1024];
-                    loop {
-                        match stderr.read(&mut buf[..]) {
-                            Err(_) | Ok(0) => {
-                                break;
-                            }
-                            Ok(n) => {
-                                log::debug!(
-                                    "Curl failed with output {}",
-                                    String::from_utf8_lossy(&buf[..n])
-                                );
-                            }
-                        }
-                    }
-                    drop(stderr);
-                }));
-                drop(std::thread::spawn(move || {
-                    let _ = child.wait();
-                }));
-            } else {
-                // Should never happen
-                let _ = control_and_errors.send(Operation::MessageDeliveryError(
-                    "Failed to get stdin/stdout/stderr".to_string(),
-                ));
-            }
-        }
-        Err(e) => {
-            // Not found, permission denied, etc.
-            let _ = control_and_errors
-                .send(Operation::Fatal(format!("Failed to launch curl: {:?}", e)));
-        }
+        (10 /* per batch */, 200 /* per message */)
     }
 }
