+++ src/datasink/mod.rs
@@ -3,6 +3,8 @@ mod background;
 pub mod delay;
 pub mod directory;
 #[cfg(feature = "kafka")]
+mod http_upload;
+#[cfg(feature = "kafka")]
 pub mod kafka;
 pub mod stdio;
